requirements.txt
matplotlib==3.8.4
python-dotenv==1.0.1
statsmodels==0.14.4
plotly-express==0.4.1
plotly==6.0.0
fastapi==0.115.12
uvicorn==0.34.2
joblib==1.5.0
numpy==1.26.4
pandas==2.2.3
requests==2.31.0
pytz==2024.2
timezonefinder==6.5.8
scikit-learn==1.6.1
xgboost==3.0.0
huggingface_hub==0.34.4
streamlit==1.47.0